[FE][Refactor] : redraw 훅 분리
